pyproject.toml
[tool.poetry]
name = "form-generator"
version = "0.1.0"
description = ""
authors = ["Sergio <sergioricjr7@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
pytest = "^8.3.3"
dependency-injector = "^4.42.0"
uvicorn = "^0.31.1"
google-auth = "^2.35.0"
requests = "^2.32.3"
odmantic = "^1.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
